stripepayment/serializer.py
from .models import Product, OrderDetail
from rest_framework import serializers


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['name', 'price']
